requirements.txt
appdirs==1.4.3
Django==1.11
django-braces==1.11.0
django-filter==1.0.2
django-oauth-toolkit==0.12.0
django-rest-framework-social-oauth2==1.0.6
djangorestframework==3.6.2
Markdown==2.6.8
oauthlib==2.0.1
packaging==16.8
psycopg2==2.7.1
PyJWT==1.5.0
pyparsing==2.2.0
python-openid==2.2.5
python-social-auth==0.3.6
pytz==2017.2
requests==2.14.2
requests-oauthlib==0.8.0
shortuuid==0.5.0
six==1.10.0
social-auth-app-django==1.2.0
social-auth-core==1.3.0
django-cors-headers
fuzzywuzzy
drfapikey
